+++ Nostrid.Photino/Misc/ClipboardService.cs
@@ -1,0 +1,10 @@
+﻿using Nostrid.Interfaces;
+using System.Threading.Tasks;
+
+internal class ClipboardService : IClipboardService
+{
+    public async Task CopyAsync(string content)
+    {
+        await TextCopy.ClipboardService.SetTextAsync(content);
+    }
+}
