+++ Nostrid.Photino/Misc/NotificationCounter.cs
@@ -1,0 +1,9 @@
+﻿using Nostrid;
+
+internal class NotificationCounter : INotificationCounter
+{
+    public void SetNotificationCount(int count)
+    {
+        // TODO: Implement this
+    }
+}
