+++ Nostrid.Photino/Nostrid.Photino.csproj
@@ -11,6 +11,7 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0" />
 		<PackageReference Include="Photino.Blazor" Version="2.5.0" />
+		<PackageReference Include="TextCopy" Version="6.2.1" />
 	</ItemGroup>
 
   <ItemGroup>
